Got rid of signed vs unsigned warnings, especially in the map editor. Also found what was making the map editor mouse move on its own: someone (in the joystick processing) was modifying the mouse_x and/or mouse_y values to "emulate" the mouse movement. I plan to rip out all joystick interface, since the editor really is not designed for joystick control at all.

+++ src/movement.cpp
@@ -42,6 +42,7 @@ static int minimize_path(AL_CONST char *, char *, size_t);
 static int search_paths(unsigned int, int *, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int, unsigned int);
 
 
+
 /*! \brief Generates the solution path.
  *
  * The only way of doing this is walking the path backwards, minimizing and
@@ -64,8 +65,8 @@ static int compose_path(AL_CONST int *map, unsigned int target_x, unsigned int t
 {
     char temp[1024];
     int index = 0;
-    int x;
-    int y;
+    unsigned int x;
+    unsigned int y;
     int value;
 
     x = target_x;
@@ -125,7 +126,7 @@ static int compose_path(AL_CONST int *map, unsigned int target_x, unsigned int t
  */
 static void copy_map(int *map)
 {
-    register int x, y;
+    register size_t x, y;
     size_t index, entity_index;
 
     for (y = 0; y < g_map.ysize; y++)
